requirement.txt
attrdict
gevent
requests
pyyaml
dnspython
